Fix bugs in get_id_correspondeces
The method did not actually work as it was untested. I've fixed it to run, but
have not yet added tests. I'm still not conviced I'm computing the new IDs
correctly so I don't have any tests. And removed all that commented out code at
the end I forgot about, whoops.

In addition I added 1D4R for tests in the future.