Handle missing assembly_gen structures.
According to John. In this case the whole AU is in one file. This means there is
no biological assemblies so the translation is pretty simple, all symmetries are
1_555. This commit creates a translation table where all access will produce a
1_555 using a defaultdict. I don't yet know where to get all models from the cif
file and I'm not sure it's worth it. In the future it may make sense to change
the program to only build a table for the valid models.