Show atom field and use | separators
We have decided to switch to using | instead of _ as a separator of the fields.
This commit uses them. In addition, we also show the atom field if needed.
Before we always left it out. We shouldn't do this and should show it when
something to the right needs to be shown.